+++ blocks/sitemap-feature-block/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @wpmedia/sitemaps-feature-block
 
+## 0.4.0-tk.0
+
+### Minor Changes
+
+- Added rss to prop-types
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @wpmedia/feeds-prop-types@0.2.0-tk.0
+
 ## 0.3.1
 
 ### Patch Changes